+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Migrated from `rye` to `uv` for development management.
+
 ## v0.5.0
 
 **Released: 2025-08-11**